pyproject.toml
[project]
name = "course-5"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirill Gladkikh",email = "k.gladkikh@mail.ru"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.17.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


# Конфигурация линтеров:
[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.mypy]
disallow_untyped_defs = true
# no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
# максимальная длина строки
line_length = 119

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]
